refactor: split large files and remove dead code for readability
Structural readability pass (behavior-preserving except one bug fix):

- Remove orphaned repo-root files (PEMKeyHelpers.swift, SampleData.swift)
  that were tracked but outside the synchronized-group target.
- Extract PEM key detection into Classes/PEMKeyDetection.swift.
- Split MainView.swift into KeyValidation, CredentialsSheetView, and
  ConnectionIndicatorView; factor duplicated notice banners and the
  create/save closure into helpers.
- Split SSHManager.swift into SSHChannelHandlers (NIO handlers) and
  SSHKeyParsing (OpenSSHKeyParser + FlexibleAuthDelegate).
- ConnectionStore: drop unused duplicate CKContainer; demote [TEMP]
  error-level logs to info.
- KeychainService: centralize account-key format in CredentialAccount,
  shared by the real and mock stores.
- NavigationView: remove dead comment and a force-unwrap.
- Fix toolbar Cancel button: clear the create form before switching to
  .view (the prior check ran after the mode change and never fired).
