+++ SSH TunnelBuilder/Classes/PEMKeyDetection.swift
@@ -3,16 +3,15 @@ import Foundation
 import AppKit
 #endif
 
-// MARK: - PEM Key Detection and Validation
+// MARK: - PEM Key Detection Utilities
 
 /// Represents the different types of PEM-encoded private keys
 enum PEMKeyKind {
-    case pkcs8        // Standard PKCS#8 format (most compatible)
-    case ec           // Elliptic Curve key (supported)
-    case rsa          // Legacy RSA format (may need conversion)
-    case openssh      // OpenSSH format (not directly supported by NIOSSH)
-    case dsa          // DSA format (deprecated, not recommended)
-    case ed25519      // ED25519 format (not yet supported)
+    case pkcs8        // PKCS#8 format - ECDSA supported (encrypted or unencrypted)
+    case ec           // EC PRIVATE KEY - ECDSA supported (unencrypted)
+    case rsa          // RSA format - NOT supported
+    case openssh      // OpenSSH format - Ed25519/ECDSA supported (unencrypted)
+    case dsa          // DSA format - NOT supported
     case unknown      // Unrecognized format
 }
 
@@ -22,16 +21,13 @@ func keyKindDescription(_ kind: PEMKeyKind) -> String {
     case .pkcs8: return "PKCS#8 PRIVATE KEY"
     case .ec: return "EC PRIVATE KEY (ECDSA)"
     case .rsa: return "RSA PRIVATE KEY"
-    case .openssh: return "OPENSSH PRIVATE KEY"
+    case .openssh: return "OPENSSH PRIVATE KEY (Ed25519/ECDSA)"
     case .dsa: return "DSA PRIVATE KEY"
-    case .ed25519: return "ED25519 PRIVATE KEY"
     case .unknown: return "Unknown"
     }
 }
 
 /// Detects the type of PEM private key from its text content
-/// - Parameter text: The PEM-encoded key text
-/// - Returns: The detected key type
 func detectPEMKeyKind(_ text: String) -> PEMKeyKind {
     let t = text.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
     if t.contains("-----BEGIN ENCRYPTED PRIVATE KEY-----") { return .pkcs8 }
@@ -40,13 +36,10 @@ func detectPEMKeyKind(_ text: String) -> PEMKeyKind {
     if t.contains("-----BEGIN RSA PRIVATE KEY-----") { return .rsa }
     if t.contains("-----BEGIN OPENSSH PRIVATE KEY-----") { return .openssh }
     if t.contains("-----BEGIN DSA PRIVATE KEY-----") { return .dsa }
-    if t.contains("ED25519 PRIVATE KEY") { return .ed25519 }
     return .unknown
 }
 
 /// Determines whether a PEM private key is encrypted
-/// - Parameter text: The PEM-encoded key text
-/// - Returns: `true` if the key is encrypted and requires a passphrase
 func isPEMEncrypted(_ text: String) -> Bool {
     let t = text.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
     if t.contains("-----BEGIN ENCRYPTED PRIVATE KEY-----") { return true }
@@ -55,8 +48,9 @@ func isPEMEncrypted(_ text: String) -> Bool {
     return false
 }
 
-// MARK: - Clipboard Helper
+// MARK: - Clipboard
 
+/// Copies text to the system clipboard (macOS only)
 func copyToClipboard(_ text: String) {
     #if os(macOS)
     let pb = NSPasteboard.general
