+++ SSH TunnelBuilder/Classes/SSHManager.swift
@@ -4,484 +4,10 @@ import NIOFoundationCompat
 @preconcurrency import NIOSSH
 import CryptoKit
 
-// NOTE: PEMKeyKind, detectPEMKeyKind, isPEMEncrypted are assumed to be defined in MainView.swift or another file.
-// If they are not visible here, move them to a shared utility file.
-
-// MARK: - Helper Types
-
-private final class InteractiveHostKeyDelegate: NIOSSHClientServerAuthenticationDelegate, Sendable {
-    let host: String
-    let port: Int
-    let validationHandler: @Sendable (String, Int, NIOSSHPublicKey, EventLoopPromise<Void>) -> Void
-
-    init(host: String, port: Int, validationHandler: @escaping @Sendable (String, Int, NIOSSHPublicKey, EventLoopPromise<Void>) -> Void) {
-        self.host = host
-        self.port = port
-        self.validationHandler = validationHandler
-    }
-
-    func validateHostKey(hostKey: NIOSSHPublicKey, validationCompletePromise: EventLoopPromise<Void>) {
-        validationHandler(host, port, hostKey, validationCompletePromise)
-    }
-}
-
-// MARK: - NIOSSH Handlers
-
-private final class SSHSessionReadyHandler: ChannelInboundHandler, RemovableChannelHandler {
-    typealias InboundIn = Any
-
-    private let sessionReadyPromise: EventLoopPromise<Void>
-    private var promiseFulfilled = false
-
-    init(sessionReadyPromise: EventLoopPromise<Void>) {
-        self.sessionReadyPromise = sessionReadyPromise
-    }
-
-    func userInboundEventTriggered(context: ChannelHandlerContext, event: Any) {
-        if event is UserAuthSuccessEvent, !promiseFulfilled {
-            promiseFulfilled = true
-            sessionReadyPromise.succeed(())
-        }
-        context.fireUserInboundEventTriggered(event)
-    }
-
-    func errorCaught(context: ChannelHandlerContext, error: Error) {
-        if !promiseFulfilled {
-            promiseFulfilled = true
-            sessionReadyPromise.fail(error)
-        }
-        context.fireErrorCaught(error)
-    }
-    
-    func channelInactive(context: ChannelHandlerContext) {
-        if !promiseFulfilled {
-            promiseFulfilled = true
-            sessionReadyPromise.fail(ChannelError.ioOnClosedChannel)
-        }
-        context.fireChannelInactive()
-    }
-}
-
-private extension TaskGroup where ChildTaskResult == Void {
-    mutating func closeChannel(_ channel: Channel?, name: String) {
-        guard let channel = channel else { return }
-        self.addTask {
-            do {
-                try await channel.close(mode: .all).get()
-            } catch {
-                Logger.error("Error closing \(name) channel: \(error)", log: Logger.ssh)
-            }
-        }
-    }
-}
-
-// `@unchecked Sendable`: a `ChannelHandler` is confined to its channel's event
-// loop, and every stored property here is immutable. NIO's pipeline APIs require
-// `Sendable` handlers under Swift 6, so we vouch for the confinement explicitly.
-// `OutboundType: Sendable` lets the transformed value cross into the peer's
-// event-loop closure without warning (the concrete types are `ByteBuffer` and
-// `SSHChannelData`, both `Sendable`).
-private final class GenericRelayHandler<InboundType, OutboundType: Sendable>: ChannelInboundHandler, @unchecked Sendable {
-    typealias InboundIn = InboundType
-
-    private let peer: Channel
-    private let onBytes: @Sendable (Int) -> Void
-    private let transform: @Sendable (InboundType) -> (OutboundType?, Int)
-
-    init(peer: Channel, onBytes: @escaping @Sendable (Int) -> Void, transform: @escaping @Sendable (InboundType) -> (OutboundType?, Int)) {
-        self.peer = peer
-        self.onBytes = onBytes
-        self.transform = transform
-    }
-
-    func channelRead(context: ChannelHandlerContext, data: NIOAny) {
-        let inboundData = self.unwrapInboundIn(data)
-        let (outboundData, count) = transform(inboundData)
-
-        if count > 0 {
-            self.onBytes(count)
-            if let outboundData {
-                // Ensure write happens on the peer's event loop for thread safety.
-                // Capture `peer` (Sendable) rather than `self`.
-                let peer = self.peer
-                peer.eventLoop.execute {
-                    peer.writeAndFlush(outboundData, promise: nil)
-                }
-            }
-        }
-    }
-
-    func channelInactive(context: ChannelHandlerContext) {
-        // Ensure close happens on the peer's event loop for thread safety.
-        let peer = self.peer
-        peer.eventLoop.execute {
-            peer.close(mode: .all, promise: nil)
-        }
-        context.fireChannelInactive()
-    }
-
-    func errorCaught(context: ChannelHandlerContext, error: Error) {
-        Logger.error("GenericRelayHandler error: \(error)", log: Logger.ssh)
-        // Ensure close happens on the peer's event loop for thread safety.
-        let peer = self.peer
-        peer.eventLoop.execute {
-            peer.close(mode: .all, promise: nil)
-        }
-        context.close(promise: nil)
-    }
-}
-
-internal enum OpenSSHKeyParser {
-    enum OpenSSHParsingError: LocalizedError, Equatable {
-        case insufficientData
-        case invalidPEMFormat
-        case invalidKeyFormat(reason: String)
-        case unsupportedKeyType(String)
-        case unsupportedCurve(String)
-        case invalidStringData
-
-        var errorDescription: String? {
-            switch self {
-            case .insufficientData:
-                return "Insufficient data while parsing the key."
-            case .invalidPEMFormat:
-                return "Invalid PEM format."
-            case .invalidKeyFormat(let reason):
-                return "Invalid key format: \(reason)"
-            case .unsupportedKeyType(let type):
-                return "Unsupported key type: '\(type)'."
-            case .unsupportedCurve(let curve):
-                return "Unsupported elliptic curve: '\(curve)'. Supported curves are nistp256, nistp384, and nistp521."
-            case .invalidStringData:
-                return "Invalid string data in key."
-            }
-        }
-    }
-
-    static func extractOpenSSHData(from pem: String) throws -> Data { return try FlexibleAuthDelegate.extractOpenSSHData(from: pem) }
-    static func parseOpenSSHPrivateKey(_ data: Data, passphrase: String? = nil) throws -> NIOSSHPrivateKey { return try FlexibleAuthDelegate.parseOpenSSHPrivateKey(data, passphrase: passphrase) }
-    static func readSSHBytes(from buffer: inout ByteBuffer) throws -> [UInt8] { return try FlexibleAuthDelegate.readSSHBytes(from: &buffer) }
-}
-
-final class FlexibleAuthDelegate: NIOSSHClientUserAuthenticationDelegate, Sendable {
-    let username: String
-    let password: String?
-    let privateKey: NIOSSHPrivateKey?
-    let initializationError: String? // Captured synchronously for immediate reporting
-    let reportError: (@Sendable (String) -> Void)?
-
-    init(username: String, password: String?, privateKeyString: String?, privateKeyPassphrase: String?, reportError: (@Sendable (String) -> Void)? = nil) {
-        self.reportError = reportError
-        self.username = username
-        self.password = password?.isEmpty == false ? password : nil
-        
-        var parsedKey: NIOSSHPrivateKey? = nil
-        var initError: String? = nil
-        
-        if let keyString = privateKeyString, !keyString.isEmpty {
-            do {
-                parsedKey = try FlexibleAuthDelegate.parsePrivateKey(pemString: keyString, passphrase: privateKeyPassphrase)
-                Logger.info("Private key parsed successfully", log: Logger.ssh)
-            } catch {
-                let errorMsg = "Failed to parse private key: \(error.localizedDescription)"
-                initError = error.localizedDescription
-                reportError?(errorMsg)
-                Logger.error(errorMsg, log: Logger.ssh)
-            }
-        }
-        self.privateKey = parsedKey
-        self.initializationError = initError
-    }
-    
-    private func makeAuthOffer(with offerType: NIOSSHUserAuthenticationOffer.Offer) -> NIOSSHUserAuthenticationOffer {
-        return NIOSSHUserAuthenticationOffer(
-            username: username,
-            serviceName: "ssh-connection",
-            offer: offerType
-        )
-    }
-
-    func nextAuthenticationType(
-        availableMethods: NIOSSHAvailableUserAuthenticationMethods,
-        nextChallengePromise: EventLoopPromise<NIOSSHUserAuthenticationOffer?>
-    ) {
-        if availableMethods.contains(.publicKey), let key = self.privateKey {
-             // Correctly construct the private key offer wrapper
-             let offer = makeAuthOffer(with: .privateKey(.init(privateKey: key)))
-             nextChallengePromise.succeed(offer)
-             return
-        }
-        
-        if availableMethods.contains(.password), let password = self.password {
-            let offer = makeAuthOffer(with: .password(.init(password: password)))
-            nextChallengePromise.succeed(offer)
-            return
-        }
-
-        nextChallengePromise.succeed(nil)
-    }
-}
-
-private extension FlexibleAuthDelegate {
-    static func parsePrivateKey(pemString: String, passphrase: String?) throws -> NIOSSHPrivateKey {
-        let trimmed = pemString.trimmingCharacters(in: .whitespacesAndNewlines)
-        
-        let kind = detectPEMKeyKind(trimmed)
-        
-        // Handle OpenSSH keys (Ed25519, ECDSA), encrypted or not
-        if kind == .openssh {
-            let data = try extractOpenSSHData(from: trimmed)
-            return try parseOpenSSHPrivateKey(data, passphrase: passphrase)
-        }
-        
-        guard kind == .pkcs8 || kind == .ec else {
-            throw SSHTunnelError.unsupportedKeyType(keyKindDescription(kind))
-        }
-
-        let isEncrypted = isPEMEncrypted(trimmed)
-        let derData: Data
-
-        if isEncrypted {
-            guard let passphrase = passphrase, !passphrase.isEmpty else {
-                throw SSHTunnelError.encryptedKeyNoPassphrase
-            }
-            derData = try PEMDecryptor.decryptEncryptedPKCS8PEM(trimmed, passphrase: passphrase)
-        } else if trimmed.contains("-----BEGIN PRIVATE KEY-----") || trimmed.contains("-----BEGIN EC PRIVATE KEY-----") {
-            derData = try decodeUnencryptedPEM(trimmed)
-        } else {
-            throw PEMDecryptorError.invalidPEM
-        }
-        
-        // A `.ec` kind here is a top-level SEC1 `EC PRIVATE KEY` (unencrypted);
-        // its DER is a bare SEC1 ECPrivateKey, not a PKCS#8 PrivateKeyInfo, so
-        // it needs the SEC1 parser. Everything else (`BEGIN PRIVATE KEY` and any
-        // decrypted PKCS#8 blob) is PKCS#8.
-        let parsedKey = (kind == .ec)
-            ? try PEMDecryptor.parseSEC1ECPrivateKey(derData)
-            : try PEMDecryptor.parsePKCS8PrivateKey(derData)
-
-        switch parsedKey {
-        case .ec(_, let privateScalar):
-            switch privateScalar.count {
-            case 32:
-                let key = try P256.Signing.PrivateKey(rawRepresentation: privateScalar)
-                return NIOSSHPrivateKey(p256Key: key)
-            case 48:
-                let key = try P384.Signing.PrivateKey(rawRepresentation: privateScalar)
-                return NIOSSHPrivateKey(p384Key: key)
-            case 66:
-                let key = try P521.Signing.PrivateKey(rawRepresentation: privateScalar)
-                return NIOSSHPrivateKey(p521Key: key)
-            default:
-                throw SSHTunnelError.unsupportedCurveLength
-            }
-        case .ed25519(let seed):
-            let key = try Curve25519.Signing.PrivateKey(rawRepresentation: seed)
-            return NIOSSHPrivateKey(ed25519Key: key)
-        case .rsa:
-            throw SSHTunnelError.rsaNotSupported
-        }
-    }
-    
-    // MARK: - OpenSSH Parsing Helpers
-    
-    static func extractOpenSSHData(from pem: String) throws -> Data {
-        let lines = pem.components(separatedBy: .newlines)
-        var base64String = ""
-        var insideBlock = false
-        
-        for line in lines {
-            if line.contains("-----BEGIN OPENSSH PRIVATE KEY-----") {
-                insideBlock = true
-                continue
-            }
-            if line.contains("-----END OPENSSH PRIVATE KEY-----") {
-                insideBlock = false
-                break
-            }
-            if insideBlock {
-                base64String += line.trimmingCharacters(in: .whitespaces)
-            }
-        }
-        
-        guard !base64String.isEmpty, let data = Data(base64Encoded: base64String) else {
-            throw OpenSSHKeyParser.OpenSSHParsingError.invalidPEMFormat
-        }
-        return data
-    }
-
-    static func parseOpenSSHPrivateKey(_ data: Data, passphrase: String? = nil) throws -> NIOSSHPrivateKey {
-        var buffer = ByteBuffer(data: data)
-
-        // 1. Magic "openssh-key-v1\0" (15 bytes)
-        guard let magic = buffer.readBytes(length: 15),
-              String(bytes: magic, encoding: .utf8) == "openssh-key-v1\0" else {
-            throw OpenSSHKeyParser.OpenSSHParsingError.invalidKeyFormat(reason: "Invalid OpenSSH magic header")
-        }
-
-        // 2-4. Cipher, KDF name, KDF options
-        let cipherName = try readSSHString(from: &buffer)
-        let kdfName = try readSSHString(from: &buffer)
-        let kdfOptions = try readSSHBytes(from: &buffer)
-        let isEncrypted = !(cipherName == "none" && kdfName == "none")
-
-        // 5. Num Keys
-        guard let numKeys = buffer.readInteger(as: UInt32.self) else {
-             throw OpenSSHKeyParser.OpenSSHParsingError.invalidKeyFormat(reason: "Could not read number of keys from private key blob")
-        }
-
-        guard numKeys >= 1 else {
-             throw OpenSSHKeyParser.OpenSSHParsingError.invalidKeyFormat(reason: "No keys found in private key data")
-        }
-
-        // 6. Public Key Blob (Binary data)
-        let _ = try readSSHBytes(from: &buffer) // pubKeyBlob
-
-        // 7. Private Key section — encrypted unless cipher/KDF are both "none".
-        // When encrypted, OpenSSH derives the cipher key+IV from the passphrase
-        // with bcrypt_pbkdf; decrypt back to the same plaintext layout. For AEAD
-        // ciphers the 16-byte authentication tag follows the encrypted section
-        // as trailing bytes, so capture whatever remains in the buffer.
-        let privateSection = try readSSHBytes(from: &buffer)
-        let authTag = buffer.readableBytes > 0
-            ? (buffer.readBytes(length: buffer.readableBytes) ?? [])
-            : []
-        let privateBlobBytes: [UInt8]
-        if isEncrypted {
-            privateBlobBytes = try OpenSSHKeyDecryptor.decryptPrivateSection(
-                cipherName: cipherName,
-                kdfName: kdfName,
-                kdfOptions: kdfOptions,
-                ciphertext: privateSection,
-                authTag: authTag,
-                passphrase: passphrase
-            )
-        } else {
-            privateBlobBytes = privateSection
-        }
-        var privateBuffer = ByteBuffer(bytes: privateBlobBytes)
-
-        // Check Ints — for an encrypted key this is also the proof that the
-        // passphrase was correct (a wrong passphrase yields garbage plaintext).
-        guard let check1 = privateBuffer.readInteger(as: UInt32.self),
-              let check2 = privateBuffer.readInteger(as: UInt32.self),
-              check1 == check2 else {
-            if isEncrypted {
-                throw OpenSSHCipherError.incorrectPassphrase
-            }
-            throw OpenSSHKeyParser.OpenSSHParsingError.invalidKeyFormat(reason: "Private key blob integrity check failed")
-        }
-        
-        // Key Type
-        let keyType = try readSSHString(from: &privateBuffer)
-        
-        if keyType == "ssh-ed25519" {
-            // Pub key
-            let _ = try readSSHBytes(from: &privateBuffer)
-            // Priv key — OpenSSH stores 64 bytes: seed (32) || public key (32).
-            // Curve25519.Signing.PrivateKey takes the 32-byte seed as rawRepresentation.
-            let keyBytes = try readSSHBytes(from: &privateBuffer)
-            guard keyBytes.count >= 32 else {
-                throw OpenSSHKeyParser.OpenSSHParsingError.invalidKeyFormat(reason: "Ed25519 private key blob too short (\(keyBytes.count) bytes)")
-            }
-            let seed = Data(keyBytes.prefix(32))
-            let ed25519Key = try Curve25519.Signing.PrivateKey(rawRepresentation: seed)
-            return NIOSSHPrivateKey(ed25519Key: ed25519Key)
-        } else if keyType.hasPrefix("ecdsa-sha2-") {
-            let curveName = try readSSHString(from: &privateBuffer)
-            let _ = try readSSHBytes(from: &privateBuffer) // Public Key
-            let privateScalarBytes = try readSSHBytes(from: &privateBuffer) // Private Key Scalar
-            
-            let normalized: Data
-            switch curveName {
-            case "nistp256":
-                normalized = try normalizeScalar(privateScalarBytes, targetSize: 32)
-                let key = try P256.Signing.PrivateKey(rawRepresentation: normalized)
-                return NIOSSHPrivateKey(p256Key: key)
-            case "nistp384":
-                normalized = try normalizeScalar(privateScalarBytes, targetSize: 48)
-                let key = try P384.Signing.PrivateKey(rawRepresentation: normalized)
-                return NIOSSHPrivateKey(p384Key: key)
-            case "nistp521":
-                normalized = try normalizeScalar(privateScalarBytes, targetSize: 66)
-                let key = try P521.Signing.PrivateKey(rawRepresentation: normalized)
-                return NIOSSHPrivateKey(p521Key: key)
-            default:
-                throw OpenSSHKeyParser.OpenSSHParsingError.unsupportedCurve(curveName)
-            }
-        } else {
-            throw OpenSSHKeyParser.OpenSSHParsingError.unsupportedKeyType(keyType)
-        }
-    }
-    
-    static func normalizeScalar(_ bytes: [UInt8], targetSize: Int) throws -> Data {
-        var result = bytes
-        // Strip leading zeros if longer (OpenSSH MPINT format)
-        while result.count > targetSize && result.first == 0 {
-            result.removeFirst()
-        }
-        
-        if result.count > targetSize {
-            throw OpenSSHKeyParser.OpenSSHParsingError.invalidKeyFormat(reason: "Private key scalar is too large for the curve")
-        }
-        
-        // Pad with leading zeros if shorter
-        while result.count < targetSize {
-            result.insert(0, at: 0)
-        }
-        
-        return Data(result)
-    }
-
-    static func readSSHString(from buffer: inout ByteBuffer) throws -> String {
-        let bytes = try readSSHBytes(from: &buffer)
-        guard let string = String(bytes: bytes, encoding: .utf8) else {
-             throw OpenSSHKeyParser.OpenSSHParsingError.invalidStringData
-        }
-        return string
-    }
-    
-    static func readSSHBytes(from buffer: inout ByteBuffer) throws -> [UInt8] {
-        guard let length = buffer.readInteger(as: UInt32.self) else {
-            throw OpenSSHKeyParser.OpenSSHParsingError.insufficientData
-        }
-        guard let bytes = buffer.readBytes(length: Int(length)) else {
-             throw OpenSSHKeyParser.OpenSSHParsingError.insufficientData
-        }
-        return bytes
-    }
-    
-    private static func decodeUnencryptedPEM(_ pem: String) throws -> Data {
-        let normalized = pem
-            .replacingOccurrences(of: "\r\n", with: "\n")
-            .replacingOccurrences(of: "\r", with: "\n")
-        
-        let startMarkers = ["-----BEGIN PRIVATE KEY-----", "-----BEGIN EC PRIVATE KEY-----"]
-        let endMarkers = ["-----END PRIVATE KEY-----", "-----END EC PRIVATE KEY-----"]
-        
-        var base64Content = ""
-        
-        for (start, end) in zip(startMarkers, endMarkers) {
-            if let startRange = normalized.range(of: start),
-               let endRange = normalized.range(of: end) {
-                let base64Start = startRange.upperBound
-                let base64End = endRange.lowerBound
-                base64Content = normalized[base64Start..<base64End]
-                    .components(separatedBy: .newlines)
-                    .joined()
-                    .trimmingCharacters(in: .whitespacesAndNewlines)
-                break
-            }
-        }
-        
-        guard !base64Content.isEmpty, let data = Data(base64Encoded: base64Content) else {
-            throw PEMDecryptorError.invalidPEM
-        }
-        return data
-    }
-}
-
-// MARK: - SSHManager
+// SSH connection lifecycle: authenticate, bring up the session, and run a local
+// listener that forwards each accepted TCP connection over a direct-tcpip SSH
+// channel. The NIO handlers it installs live in `SSHChannelHandlers.swift`, and
+// private-key parsing lives in `SSHKeyParsing.swift`.
 
 final class SSHManager: @unchecked Sendable {
     /// Connection timeout in seconds. Increase for high-latency networks.
@@ -559,7 +85,7 @@ final class SSHManager: @unchecked Sendable {
                 }
             )
         }.value
-        
+
         // Check for initialization errors first - these are fatal
         if let initError = authDelegate.initializationError {
             let errorMsg = "Failed to parse private key: \(initError)"
@@ -583,7 +109,7 @@ final class SSHManager: @unchecked Sendable {
 
             throw SSHTunnelError.keyParsingFailed(detail)
         }
-        
+
         let host = connectionInfo.serverAddress
         let port = Int(connectionInfo.portNumber) ?? 22
 
@@ -620,12 +146,12 @@ final class SSHManager: @unchecked Sendable {
 
         do {
             let channel = try await bootstrap.connect(host: host, port: port).get()
-            
+
             if let sessionReadyPromise = self.sessionReadyPromise {
                 try await sessionReadyPromise.futureResult.get()
                 self.sessionReadyCompleted = true
             }
-            
+
             lock.withLock { self.sshClientChannel = channel }
             await MainActor.run {
                 self.connection.state = .connected
@@ -648,7 +174,9 @@ final class SSHManager: @unchecked Sendable {
             throw error
         }
     }
-    
+
+    // MARK: - Host Key Validation
+
     private func handleHostKeyValidation(host: String, port: Int, key: NIOSSHPublicKey, promise: EventLoopPromise<Void>, knownHostKey: String) {
         let keyData = serialize(key: key)
         let fingerprint: String
@@ -659,7 +187,7 @@ final class SSHManager: @unchecked Sendable {
             // Fallback: we cannot serialize the key with this NIOSSH version; present a legacy marker.
             fingerprint = "SHA256:UNAVAILABLE"
         }
-        
+
         if !knownHostKey.isEmpty {
             // Try raw key base64 match first (if we have raw key bytes)
             if let keyData = keyData, let knownData = Data(base64Encoded: knownHostKey), knownData == keyData {
@@ -692,7 +220,7 @@ final class SSHManager: @unchecked Sendable {
             allowed ? promise.succeed(()) : promise.fail(SSHTunnelError.hostKeyRejected)
         }
     }
-    
+
     private func serialize(key _: NIOSSHPublicKey) -> Data? {
         // NIOSSH does not currently expose a public API to serialize host keys to raw bytes.
         // Returning nil causes handleHostKeyValidation to use fingerprint-based verification,
@@ -700,6 +228,8 @@ final class SSHManager: @unchecked Sendable {
         // If NIOSSH adds serialization support in the future, implement it here for raw key matching.
         return nil
     }
+
+    // MARK: - Local Listener / Forwarding
 
     private func startLocalListener() async throws {
         // Capture active tunnel info safely
@@ -729,20 +259,20 @@ final class SSHManager: @unchecked Sendable {
                     sshChannel.pipeline.handler(type: NIOSSHHandler.self).flatMap { sshHandler -> EventLoopFuture<Channel> in
                         let childChannelPromise = sshChannel.eventLoop.makePromise(of: Channel.self)
                         Logger.debug("Opening direct-tcpip to \(tunnelInfo.remoteServer):\(Int(tunnelInfo.remotePort) ?? 0) from originator: \(originator)", log: Logger.ssh)
-                        
+
                         let directTCPIP = NIOSSH.SSHChannelType.DirectTCPIP(
                             targetHost: tunnelInfo.remoteServer,
                             targetPort: Int(tunnelInfo.remotePort) ?? 0,
                             originatorAddress: originator
                         )
-                        
+
                         sshHandler.createChannel(childChannelPromise, channelType: .directTCPIP(directTCPIP)) { sshChildChannel, _ in
                             strongSelf.configureSSHChildPipeline(sshChildChannel, inbound: inbound, strongSelf: strongSelf)
                         }
                         return childChannelPromise.futureResult
                     }
                 }
-                
+
                 let setupFuture = inbound.setOption(ChannelOptions.autoRead, value: false).flatMap {
                     sshChildChannelFuture
                 }.flatMap { sshChildChannel -> EventLoopFuture<Void> in
@@ -759,7 +289,7 @@ final class SSHManager: @unchecked Sendable {
                     }
                     inbound.close(promise: nil)
                 }
-                
+
                 return setupFuture
             }
             .childChannelOption(ChannelOptions.socketOption(.so_reuseaddr), value: 1)
@@ -781,8 +311,8 @@ final class SSHManager: @unchecked Sendable {
             throw error
         }
     }
-    
-    // New helper method extracted from the inner closure in startLocalListener()
+
+    /// Relays bytes arriving on the SSH child channel out to the local TCP peer.
     private func configureSSHChildPipeline(_ sshChildChannel: Channel, inbound: Channel, strongSelf: SSHManager) -> EventLoopFuture<Void> {
         let received: @Sendable (Int) -> Void = { [weak strongSelf] n in
             guard let manager = strongSelf else { return }
@@ -795,6 +325,7 @@ final class SSHManager: @unchecked Sendable {
         return sshChildChannel.pipeline.addHandler(sshToTCP)
     }
 
+    /// Relays bytes arriving on the local TCP peer into the SSH child channel.
     private func configureInboundTCPPipeline(_ inbound: Channel, sshChildChannel: Channel, strongSelf: SSHManager) -> EventLoopFuture<Void> {
         let sent: @Sendable (Int) -> Void = { [weak strongSelf] n in
             guard let manager = strongSelf else { return }
@@ -806,6 +337,8 @@ final class SSHManager: @unchecked Sendable {
         }
         return inbound.pipeline.addHandler(tcpToSSH)
     }
+
+    // MARK: - Disconnect / Shutdown
 
     func disconnect() async {
         await MainActor.run {
@@ -854,9 +387,9 @@ final class SSHManager: @unchecked Sendable {
             self.connection.bytesReceived = 0
         }
     }
-    
+
     // MARK: - Main Actor State Updates
-    
+
     private func handleBytesSent(_ count: Int) {
         let n = Int64(count)
         Task { @MainActor in
@@ -871,4 +404,3 @@ final class SSHManager: @unchecked Sendable {
         }
     }
 }
-
